requirements.txt
fastapi
uvicorn
bcrypt
cryptography
sqlalchemy
python-dotenv
slowapi
